refactor : separate components
